@@ -2,6 +2,7 @@ use axum::routing::get;
 use axum::{http, Router};
 use clap::Args;
 
+use mono::context::AppContext;
 use quinn::rustls;
 use tower::ServiceBuilder;
 use tower_http::cors::{Any, CorsLayer};
@@ -9,7 +10,7 @@ use tower_http::decompression::RequestDecompressionLayer;
 use tower_http::trace::TraceLayer;
 
 use common::model::{CommonHttpOptions, P2pOptions};
-use jupiter::context::Context;
+use jupiter::context::Storage;
 use mono::api::lfs::lfs_router;
 use mono::api::MonoApiServiceState;
 use mono::server::https_server::{get_method_router, post_method_router, AppState};
@@ -27,7 +28,7 @@ pub struct HttpOptions {
     pub p2p: P2pOptions,
 }
 
-pub async fn http_server(context: Context, options: HttpOptions) {
+pub async fn http_server(context: AppContext, options: HttpOptions) {
     let HttpOptions {
         common: CommonHttpOptions { host, port, .. },
         p2p,
@@ -39,7 +40,7 @@ pub async fn http_server(context: Context, options: HttpOptions) {
 
     check_run_with_p2p(context.clone(), options.p2p.clone());
 
-    let app = app(context, host.clone(), port, p2p.clone()).await;
+    let app = app(context.storage, host.clone(), port, p2p.clone()).await;
 
     let server_url = format!("{}:{}", host, port);
 
@@ -49,15 +50,15 @@ pub async fn http_server(context: Context, options: HttpOptions) {
         .unwrap();
 }
 
-pub async fn app(context: Context, host: String, port: u16, p2p: P2pOptions) -> Router {
+pub async fn app(storage: Storage, host: String, port: u16, p2p: P2pOptions) -> Router {
     let state = AppState {
         host: host.clone(),
         port,
-        context: context.clone(),
+        storage: storage.clone(),
     };
 
     let mono_api_state = MonoApiServiceState {
-        context: context.clone(),
+        storage: storage.clone(),
         oauth_client: None,
         store: None,
         listen_addr: format!("http://{}:{}", host, port),
@@ -103,7 +104,7 @@ pub async fn app(context: Context, host: String, port: u16, p2p: P2pOptions) -> 
     router
 }
 
-pub fn check_run_with_p2p(context: Context, p2p: P2pOptions) {
+pub fn check_run_with_p2p(context: AppContext, p2p: P2pOptions) {
     //Mega server join a ztm mesh
     match p2p.bootstrap_node {
         Some(bootstrap_node) => {
@@ -113,7 +114,9 @@ pub fn check_run_with_p2p(context: Context, p2p: P2pOptions) {
             );
 
             tokio::spawn(async move {
-                if let Err(e) = gemini::p2p::client::run(context, bootstrap_node).await {
+                if let Err(e) =
+                    gemini::p2p::client::run(context.storage, context.vault, bootstrap_node).await
+                {
                     tracing::error!("P2P client closed:{}", e)
                 }
             });
