@@ -22,9 +22,9 @@ pub struct GitEventReq {
 }
 
 impl GitEventReq {
-    pub async fn to_git_event(&self, identifier: String, commit: String) -> GitEvent {
+    pub async fn to_git_event(&self, peer_id: String, identifier: String, commit: String) -> GitEvent {
         GitEvent {
-            peer: vault::get_peerid().await,
+            peer: peer_id,
             uri: identifier,
             action: self.action.clone(),
             r#ref: "".to_string(),
