@@ -288,7 +288,7 @@ async fn path_can_be_cloned(
     state: State<MonoApiServiceState>,
 ) -> Result<Json<CommonResult<bool>>, ApiError> {
     let path: PathBuf = query.path.clone().into();
-    let import_dir = state.storage.config.monorepo.import_dir.clone();
+    let import_dir = state.storage.config().monorepo.import_dir.clone();
     let res = if path.starts_with(&import_dir) {
         state
             .storage
