@@ -1,6 +1,6 @@
+pub mod context;
 pub mod api;
 pub mod cli;
 mod commands;
-pub mod context;
 pub mod git_protocol;
 pub mod server;
